% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_area.R
\name{convert_area}
\alias{convert_area}
\title{Convert Area}
\usage{
convert_area(value, from = c("acres", "hectares"))
}
\arguments{
\item{value}{A numeric value}

\item{from}{What you're converting from}
}
\value{
A numeric value
}
\description{
Convert Area
}
\examples{
convert_area(15, from = "acres")
convert_area(15, from = "hectares")
}
